Move top of DOS dashboard into own module.
