+++ client/src/component/sos/DOSDashboard/Page.tsx
@@ -6,6 +6,8 @@ import { Tooltip } from '@blueprintjs/core';
 
 import SoSNav from '../Nav';
 
+import Main from './Main';
+
 import counties from '../../../data/counties';
 
 
@@ -152,37 +154,6 @@ const CountyUpdates = ({ countyStatus }: any) => {
     );
 };
 
-const RiskLimitInfo = ({ riskLimit }: any) => {
-    return (
-        <div className='pt-card'>
-            <strong>Target Risk Limit: </strong> { riskLimit * 100 } %
-        </div>
-    );
-};
-
-const SeedInfo = ({ seed }: any) => {
-    return (
-        <div className='pt-card'>
-            <strong>Seed: </strong> { seed }
-        </div>
-    );
-};
-
-const AuditParams = ({ sos }: any) => {
-    return (
-        <div>
-            <RiskLimitInfo riskLimit={ sos.riskLimit } />
-            <SeedInfo seed={ sos.seed } />
-        </div>
-    );
-};
-
-const AuditNotDefined = () => {
-    return (
-        <div><em>The audit has not yet been defined.</em></div>
-    );
-};
-
 const DOSDashboardPage = (props: any) => {
     const { contests, countyStatus, currentAsmState, seed, sos } = props;
 
@@ -194,16 +165,10 @@ const DOSDashboardPage = (props: any) => {
     ];
     const showAuditParams = _.includes(ongoingStates, currentAsmState);
 
-    const auditParams = showAuditParams
-                      ? <AuditParams sos={ sos } />
-                      : <AuditNotDefined />;
-
     return (
         <div className='sos-home'>
             <SoSNav />
-            <div className='sos-notifications pt-card'>
-                { auditParams }
-            </div>
+            <Main showAuditParams={ showAuditParams } sos={ sos } />
             <div className='sos-info pt-card'>
                 <CountyUpdates countyStatus={ countyStatus } />
                 <ContestUpdates contests={ contests } seed={ seed } sos={ sos } />
